# Remove Quantity.s to keep the immutability of Quantity.