--- a/staging/src/k8s.io/apimachinery/pkg/api/resource/quantity_proto_test.go
+++ b/staging/src/k8s.io/apimachinery/pkg/api/resource/quantity_proto_test.go
@@ -28,10 +28,10 @@ func TestQuantityProtoMarshal(t *testing.T) {
 		quantity string
 		expect   Quantity
 	}{
-		{"0", Quantity{i: int64Amount{value: 0, scale: 0}, s: "0", Format: DecimalSI}},
-		{"100m", Quantity{i: int64Amount{value: 100, scale: -3}, s: "100m", Format: DecimalSI}},
-		{"50m", Quantity{i: int64Amount{value: 50, scale: -3}, s: "50m", Format: DecimalSI}},
-		{"10000T", Quantity{i: int64Amount{value: 10000, scale: 12}, s: "10000T", Format: DecimalSI}},
+		{"0", Quantity{i: int64Amount{value: 0, scale: 0}, Format: DecimalSI}},
+		{"100m", Quantity{i: int64Amount{value: 100, scale: -3}, Format: DecimalSI}},
+		{"50m", Quantity{i: int64Amount{value: 50, scale: -3}, Format: DecimalSI}},
+		{"10000T", Quantity{i: int64Amount{value: 10000, scale: 12}, Format: DecimalSI}},
 	}
 	for _, testCase := range table {
 		q := MustParse(testCase.quantity)
@@ -47,9 +47,9 @@ func TestQuantityProtoMarshal(t *testing.T) {
 		dec    *inf.Dec
 		expect Quantity
 	}{
-		{dec(0, 0).Dec, Quantity{i: int64Amount{value: 0, scale: 0}, d: infDecAmount{dec(0, 0).Dec}, s: "0", Format: DecimalSI}},
-		{dec(10, 0).Dec, Quantity{i: int64Amount{value: 0, scale: 0}, d: infDecAmount{dec(10, 0).Dec}, s: "10", Format: DecimalSI}},
-		{dec(-10, 0).Dec, Quantity{i: int64Amount{value: 0, scale: 0}, d: infDecAmount{dec(-10, 0).Dec}, s: "-10", Format: DecimalSI}},
+		{dec(0, 0).Dec, Quantity{i: int64Amount{value: 0, scale: 0}, d: infDecAmount{dec(0, 0).Dec}, Format: DecimalSI}},
+		{dec(10, 0).Dec, Quantity{i: int64Amount{value: 0, scale: 0}, d: infDecAmount{dec(10, 0).Dec}, Format: DecimalSI}},
+		{dec(-10, 0).Dec, Quantity{i: int64Amount{value: 0, scale: 0}, d: infDecAmount{dec(-10, 0).Dec}, Format: DecimalSI}},
 	}
 	for _, testCase := range table2 {
 		q := Quantity{d: infDecAmount{testCase.dec}, Format: DecimalSI}
@@ -68,10 +68,10 @@ func TestQuantityProtoUnmarshal(t *testing.T) {
 		input  Quantity
 		expect string
 	}{
-		{Quantity{i: int64Amount{value: 0, scale: 0}, s: "0", Format: DecimalSI}, "0"},
-		{Quantity{i: int64Amount{value: 100, scale: -3}, s: "100m", Format: DecimalSI}, "100m"},
-		{Quantity{i: int64Amount{value: 50, scale: -3}, s: "50m", Format: DecimalSI}, "50m"},
-		{Quantity{i: int64Amount{value: 10000, scale: 12}, s: "10000T", Format: DecimalSI}, "10000T"},
+		{Quantity{i: int64Amount{value: 0, scale: 0}, Format: DecimalSI}, "0"},
+		{Quantity{i: int64Amount{value: 100, scale: -3}, Format: DecimalSI}, "100m"},
+		{Quantity{i: int64Amount{value: 50, scale: -3}, Format: DecimalSI}, "50m"},
+		{Quantity{i: int64Amount{value: 10000, scale: 12}, Format: DecimalSI}, "10000T"},
 	}
 	for _, testCase := range table {
 		var inputQ Quantity
@@ -87,9 +87,9 @@ func TestQuantityProtoUnmarshal(t *testing.T) {
 		input  Quantity
 		expect *inf.Dec
 	}{
-		{Quantity{i: int64Amount{value: 0, scale: 0}, d: infDecAmount{dec(0, 0).Dec}, s: "0", Format: DecimalSI}, dec(0, 0).Dec},
-		{Quantity{i: int64Amount{value: 0, scale: 0}, d: infDecAmount{dec(10, 0).Dec}, s: "10", Format: DecimalSI}, dec(10, 0).Dec},
-		{Quantity{i: int64Amount{value: 0, scale: 0}, d: infDecAmount{dec(-10, 0).Dec}, s: "-10", Format: DecimalSI}, dec(-10, 0).Dec},
+		{Quantity{i: int64Amount{value: 0, scale: 0}, d: infDecAmount{dec(0, 0).Dec}, Format: DecimalSI}, dec(0, 0).Dec},
+		{Quantity{i: int64Amount{value: 0, scale: 0}, d: infDecAmount{dec(10, 0).Dec}, Format: DecimalSI}, dec(10, 0).Dec},
+		{Quantity{i: int64Amount{value: 0, scale: 0}, d: infDecAmount{dec(-10, 0).Dec}, Format: DecimalSI}, dec(-10, 0).Dec},
 	}
 	for _, testCase := range table2 {
 		var inputQ Quantity