@RainbowMango 2019-09-03
@kdada  Thanks for your contribution.
Can you explain more about why we need this change?

> Quantity should not change in a Get method.

Can you provide more details about how to reproduce it? And what's the bad effect?

```go
// String formats the Quantity as a string, caching the result if not calculated.
// String is an expensive operation and caching this result significantly reduces the cost of
// normal parse / marshal operations on Quantity.
func (q *Quantity) String() string {
```
According to the comments, we cache string for improving performance, we need more reasonable reason if we want to remove the cache.