@kdada 2019-09-03
@RainbowMango 

I'm using `kubebuilder` to write controllers. In the `Reconcile` method, I want to create a merge patch if some changes are applied to the target object.  The pseudo code:
```
func Reconcile(object runtime.Object) {
    origin := object.DeepCopyObject()
    // Phase A: Do something.
    if reflect.DeepCopy(origin, object) {
        // Phase B: Create a merge patch and post to apiserver.
    }
}
``` 

I trigged a bug that it created too many patch requests to apiserver even though the method changed nothing. I found that `*resource.Quantity` will be modified when I log it. It's very confused.