@RainbowMango 2019-09-03
@kdada 
I got it.  But, just a guess, there may be other ways to solve your problem.

In another hand, even if we decide to remove the fields, we should consider more about the impact. 

Can @smarterclayton share comments?