@smarterclayton 2019-09-03
/hold

This allocation was a significant point previously.  The microbenchmark for precalc covers a case that dominated deserialization of pods.

I'd like to see alternative fixes for your use case proposed.  DeepEqual is not a sufficient goal to justify the regression.